Extract layout image request into separate file. Do not block the document load event.
