+++ components/script/lib.rs
@@ -110,6 +110,7 @@ pub mod document_loader;
 #[macro_use]
 mod dom;
 pub mod fetch;
+mod layout_image;
 pub mod layout_wrapper;
 mod mem;
 mod microtask;
